[bug] Fix 2 lock issues if db file does not exist.
Issue 1 :

the openfile is in the while. So we open the file and never close it each loop

```
In [4]: f = open('/etc/foo', 'r')

In [5]: id(f)
Out[5]: 140571115440864

In [6]: f = open('/etc/foo', 'r')

In [7]: id(f)
Out[7]: 140571115441008
```

Issue 2 :

The `open` used in the same try than the `flock`. Unfortunately, both raise the same Exception.
In case the file does not exist we never know and pastefile simply timeout on the lock.