fix #45 Change Ls behavior when file have no info
With tests.

Also fix an issue in infos_file when a file in db are not present on disk.

Without this patch, if a file is not present on disk, ls function will return a trace.

Is a file is not on disk and present in db, this is handled by clean_file
https://github.com/guits/pastefile/blob/master/pastefile/app.py#L78-L81

It's just a question of time to delete this entry in the db. `ls` should not fail
because of that.