Split upload_file into functions. And reduce actions during the lock
Create `add_new_file` and `utils.write_tmpfile_to_disk` to reduce the amount of code
in `upload_file`.

Reduce things in lock. For exemple `os.remove(tmp_full_filename)` is not in `with` statement

Rethink to simplify `upload_file`, it's not more simple to read and understand what is going on.